+++ ckGraphicsEngine/FX2dGraphicsEngine.java
@@ -780,8 +780,8 @@ CKGraphicsEngine,CKWorkSupervisorListener<CKGraphicsScene>
 
 
 	@Override
-	public void selectArea(CKPosition originLocation, float minDistance,
-			float maxDistance, CKSelectedPositionsListeners callSOUTHEAST,
+	public void selectArea(CKPosition originLocation, double minDistance,
+			double maxDistance, CKSelectedPositionsListeners callSOUTHEAST,
 			SelectAreaType type)
 	{
 		//start new transaction
@@ -800,8 +800,8 @@ CKGraphicsEngine,CKWorkSupervisorListener<CKGraphicsScene>
 
 
 	@Override
-	public void selectAreaOffsets(CKPosition originLocation, float minDistance,
-			float maxDistance, CKSelectedPositionsListeners callback,
+	public void selectAreaOffsets(CKPosition originLocation, double minDistance,
+			double maxDistance, CKSelectedPositionsListeners callback,
 			Collection<CKPosition> offsets)
 	{
 		//start new transaction
